+++ packages/terra-form-select/src/terra-dev-site/test/form-select/MaxSelection.test.jsx
@@ -14,14 +14,14 @@ const MaxSelectionExample = () => (
     <SelectField.OptGroup label="Shades of Blue">
       <SelectField.Option value="blue" display="Blue" />
       <SelectField.Option value="cyan" display="Cyan" />
-      <SelectField.Option disabled={true} value="teal" display="Teal" />
+      <SelectField.Option disabled value="teal" display="Teal" />
       <SelectField.Option value="azul" display="Azul" />
       <SelectField.Option value="aqua" display="Aqua" />
     </SelectField.OptGroup>
     <SelectField.OptGroup label="Shades of Green">
       <SelectField.Option value="green" display="Green" />
       <SelectField.Option value="forest" display="Forest Green" />
-      <SelectField.Option disabled={true} value="dark" display="Dark Green" />
+      <SelectField.Option disabled value="dark" display="Dark Green" />
       <SelectField.Option value="neon" display="Neon Green" />
       <SelectField.Option value="emerald" display="Emerald" />
     </SelectField.OptGroup>